change config helper module of test
to avoid duplicate name
