+++ test/minitest_helper.rb
@@ -1,12 +1,16 @@
 require "bundler"
 Bundler.require
 
+ENV['RACK_ENV'] = 'test'
+
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require 'kakine'
 require 'minitest/autorun'
 require 'mocha/mini_test'
 require 'fog/openstack/models/network/security_group'
 require 'fog/openstack/models/network/security_group_rule'
+# require 'yao/resources/security_group'
+# require 'yao/resources/security_group_rule'
 
 class Dummy
   def id
