+++ test/test_kakine_builder.rb
@@ -1,9 +1,9 @@
 require 'minitest_helper'
-require 'support/config_helper'
+require 'support/test_helper'
 
 class TestKakineBuilder < Minitest::Test
   def setup
-    Kakine::Resource.get(:openstack).stubs(:security_groups_hash).returns([Kakine::Config::Helper.full_rule_security_group])
+    Kakine::Resource.get(:openstack).stubs(:security_groups_hash).returns([Kakine::TestHelper.full_rule_security_group])
     Kakine::Option.stubs(:dryrun?).returns(true)
     Kakine::Option.stubs(:tenant_name).returns("test_tenant")
     Kakine::Resource.get(:openstack).stubs(:tenant).returns(Dummy.new)
@@ -11,25 +11,25 @@ class TestKakineBuilder < Minitest::Test
   end
 
   def test_create_security_group
-    full_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.full_rule_security_group) 
+    full_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.full_rule_security_group)
     assert_equal(Kakine::Builder.create_security_group(full_sg),"Create Security Group: test_full_group")
-    
-    short_rule_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.short_rule_security_group) 
+
+    short_rule_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.short_rule_security_group)
     assert_equal(Kakine::Builder.create_security_group(short_rule_sg),"Create Security Group: test_short_group")
   end
-  
+
   def test_delete_security_group
-    sg = Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.full_rule_security_group) 
+    sg = Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.full_rule_security_group)
     assert_equal(Kakine::Builder.delete_security_group(sg),"Delete Security Group: test_id_1")
   end
 
   def test_create_security_rule
-    rule = Kakine::SecurityRule.new(Kakine::Config::Helper.full_rule_port_remote_ip, "test_rule", "test_tenant") 
+    rule = Kakine::SecurityRule.new(Kakine::TestHelper.full_rule_port_remote_ip, "test_rule", "test_tenant")
     assert_equal(Kakine::Builder.create_security_rule("test_tenant", "test_security_group", rule), "Create Rule: test_security_group")
   end
 
   def test_delete_security_rule
-    rule = Kakine::SecurityRule.new(Kakine::Config::Helper.full_rule_port_remote_ip, "test_rule", "test_tenant") 
+    rule = Kakine::SecurityRule.new(Kakine::TestHelper.full_rule_port_remote_ip, "test_rule", "test_tenant")
     assert_equal(Kakine::Builder.delete_security_rule("test_tenant", "test_security_group", rule), "Delete Rule: test_id_1")
   end
 
@@ -40,34 +40,34 @@ class TestKakineBuilder < Minitest::Test
   def test_clean_up_security_group
     current_sgs = []
     new_sgs = []
-    current_sgs << Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.full_rule_security_group)
-    current_sgs << Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.short_rule_security_group)
-    new_sgs << Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.full_rule_security_group)
-    assert_equal(Kakine::Builder.clean_up_security_group(new_sgs, current_sgs), [nil, "Delete Security Group: test_id_2"]) 
+    current_sgs << Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.full_rule_security_group)
+    current_sgs << Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.short_rule_security_group)
+    new_sgs << Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.full_rule_security_group)
+    assert_equal(Kakine::Builder.clean_up_security_group(new_sgs, current_sgs), [nil, "Delete Security Group: test_id_2"])
   end
-  
+
   def test_convergence_security_group
-    current_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.full_rule_security_group)
-    lost_desc_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.lost_column("description"))
-    assert_equal(Kakine::Builder.convergence_security_group(lost_desc_sg, current_sg), ["Create Rule: loss_description_group"] ) 
-    
-    icmp_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.full_security_group(Kakine::Config::Helper.full_rule_icmp_remote_group, "icmp_group"))
-    assert_equal(Kakine::Builder.convergence_security_group(icmp_sg, current_sg), ["Create Rule: icmp_group"]) 
+    current_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.full_rule_security_group)
+    lost_desc_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.lost_column("description"))
+    assert_equal(Kakine::Builder.convergence_security_group(lost_desc_sg, current_sg), ["Create Rule: loss_description_group"] )
+
+    icmp_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.full_security_group(Kakine::TestHelper.full_rule_icmp_remote_group, "icmp_group"))
+    assert_equal(Kakine::Builder.convergence_security_group(icmp_sg, current_sg), ["Create Rule: icmp_group"])
   end
 
   def test_already_setup_security_group
     current_sgs = []
-    current_sgs << Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.full_rule_security_group)
-    current_sgs << Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.short_rule_security_group)
-    new_sg =  Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.full_rule_security_group)
+    current_sgs << Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.full_rule_security_group)
+    current_sgs << Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.short_rule_security_group)
+    new_sg =  Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.full_rule_security_group)
 
     assert_equal(Kakine::Builder.already_setup_security_group(new_sg, current_sgs).name, "test_full_group")
   end
 
   def test_clean_up_security_rule
-    current_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.full_rule_security_group)
-    new_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::Config::Helper.short_rule_security_group)
-    assert_equal(Kakine::Builder.clean_up_security_rule(new_sg, current_sg), ["Delete Rule: test_id_1"]) 
+    current_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.full_rule_security_group)
+    new_sg = Kakine::SecurityGroup.new("test_tenant", Kakine::TestHelper.short_rule_security_group)
+    assert_equal(Kakine::Builder.clean_up_security_rule(new_sg, current_sg), ["Delete Rule: test_id_1"])
   end
 
   def test_security_groups
